restructure store helpers
